pyproject.toml
[tool.poetry]
name = "shop-be"
version = "0.1.0"
description = ""
authors = ["Serhii <seruygrid@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
sqlalchemy = "^2.0.20"
uvicorn = "^0.23.2"
pydantic-settings = "^2.0.3"
asyncpg = "^0.28.0"
uvloop = "^0.17.0"
psycopg2-binary = "^2.9.8"
httpx = "^0.25.0"
ujson = "^5.8.0"
pyjwt = "^2.8.0"
aioboto3 = "^12.0.0"
pydantic = {extras = ["email"], version = "^2.5.0"}
awesome-slugify = "^1.6.5"
python-multipart = "^0.0.6"
cryptography = "^41.0.5"
python-json-logger = "^2.0.7"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
